Move Evm DB code into a submodule
